VIH-9051 - duplicate code refactor
