+++ VideoWeb/VideoWeb/ClientApp/src/app/security/conference.guard.ts
@@ -7,7 +7,6 @@ import { AuthBaseGuard } from './auth-base.guard';
 import { SecurityServiceProvider } from './authentication/security-provider.service';
 import { FeatureFlagService } from '../services/feature-flag.service';
 import { VideoWebService } from '../services/api/video-web.service';
-import { Hearing } from '../shared/models/hearing';
 
 @Injectable({
     providedIn: 'root'
@@ -28,28 +27,7 @@ export class ConferenceGuard extends AuthBaseGuard implements CanActivate {
             .pipe(take(1))
             .toPromise()
             .then(async (auth: boolean) => {
-                if (!auth) {
-                    this.router.navigate([pageUrls.Login]);
-                    return false;
-                }
-                const conferenceId = next.paramMap.get('conferenceId');
-                this.logger.debug(`[ConferenceGuard] Checking if user can view conference ${conferenceId}`);
-                try {
-                    const data = await this.videoWebService.getConferenceById(conferenceId);
-                    const hearing = new Hearing(data);
-                    if (hearing.isPastClosedTime()) {
-                        this.logger.info(
-                            '[ConferenceGuard] Returning back to hearing list because hearing has been closed for over 2 hours.'
-                        );
-                        this.router.navigate([pageUrls.JudgeHearingList]);
-                        return false;
-                    }
-                    return true;
-                } catch (err) {
-                    this.logger.error(`[ConferenceGuard] Could not get conference data. Returning home.`, err);
-                    this.router.navigate([pageUrls.Home]);
-                    return false;
-                }
+                return await this.canConferenceBeActivate(next, auth, this.videoWebService, pageUrls.JudgeHearingList, pageUrls.Home, '[ConferenceGuard]');
             });
     }
 }
